backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
pandas==2.1.3
django-cors-headers==4.3.1
reportlab==4.0.7
